pyproject.toml
[tool.poetry]
name = "lms-rest-api"
version = "0.1.0"
description = ""
authors = ["Maxim Donchuk <m.donchuk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0"
djangorestframework = "^3.14.0"
pillow = "^10.1.0"
psycopg2 = "^2.9.9"
django-environ = "^0.11.2"
ipython = "^8.18.1"
django-filter = "^23.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
